Incorporate PR feedback
(cherry picked from commit 1b3ce913f2bb1ddcfb9be047d6f860751e1efd5d)

Conflicts:
	src/ExpressionEvaluator/CSharp/Test/ExpressionCompiler/ExpressionCompilerTests.cs
	src/ExpressionEvaluator/VisualBasic/Test/ExpressionCompiler/ExpressionCompilerTests.vb
