+++ src/ExpressionEvaluator/VisualBasic/Test/ExpressionCompiler/BasicExpressionCompilerTest.vbproj
@@ -101,6 +101,7 @@
     <Compile Include="InstructionDecoderTests.vb" />
     <Compile Include="HoistedStateMachineLocalTests.vb" />
     <Compile Include="HoistedMeTests.vb" />
+    <Compile Include="ReferencedModulesTests.vb" />
     <Compile Include="ResultPropertiesTests.vb" />
     <Compile Include="NoPIATests.vb" />
     <Compile Include="PseudoVariableTests.vb" />
